Implement write_h1_summary and insert it in the training session.
write_embeddings_summary has been substituted by write_h1_summary.
The new function write inputs and outputs arithmetic operations and their number of carries.
Plus, the function record h1 that varies depending on inputs.
The outputs and number of carries are intended to be used as labels of the inputs.
Later, the recorded data will be used for visulazation.